Resolve configdir so that it's not a symlink when is_dir() is called
There are some use cases where a user might have to make ~/.config a symlink
(e.g. if a corporate home directory has limited space, and symlinks are
used to offload the storage to a separate storage disk). So resolving
configdir beforehand should prevent the is_dir() check from failing,
causing a warning to be issued and forcing a temporary directory to be
used.